+++ src/UI/ObjectEditorPopup.cpp
@@ -5,6 +5,16 @@
 
 using namespace geode::prelude;
 
+template<class OkType, class ErrType>
+void checkResult(const geode::Result<OkType, ErrType>& res, std::string& errString, const std::string& fieldName) {
+    if(res.isErr()) {
+        errString += fieldName + " Error: " + res.err().value() + '\n';
+    } else {
+        if (std::isnan(res.unwrap())) errString += fieldName + " Error: NaN.\n";
+        else if (std::isinf(res.unwrap())) errString += fieldName + " Error: Infinite.\n";
+    }
+}
+
 CCNode* ObjectEditorPopup::createNode(const char* labelText, const std::string& textInputMessage, const std::string& textInputContent, geode::TextInput*& textInput) { //*& looks awful
     CCNode* node = CCNode::create();
     node->setContentSize({100.f, 60.f});
@@ -35,56 +45,58 @@ void ObjectEditorPopup::onApplyButton(CCObject* sender) {
     auto zOrderStr = zOrderInput->getString();
     auto layerStr = editorLayerInput->getString();
 
-    float xPos = 0.f;
-    float yPos = 0.f;
+    double xPos = 0.f;
+    double yPos = 0.f;
     float xScale = 0.f;
     float yScale = 0.f;
     float rotation = 0.f;
     int zOrder = 0.f;
-    int layer = 0;
+    short layer = 0;
 
-    try {
-        xPos = std::stof(xPosStr);
-        yPos = std::stof(yPosStr);
-        xScale = std::stof(xScaleStr);
-        yScale = std::stof(yScaleStr);
-        rotation = std::stof(rotationStr);
-        zOrder = std::stoi(zOrderStr);
-        layer = std::stoi(layerStr);
+    auto xPosResult = geode::utils::numFromString<double>(xPosStr);
+    auto yPosResult = geode::utils::numFromString<double>(yPosStr);
+    auto xScaleResult = geode::utils::numFromString<float>(xScaleStr);
+    auto yScaleResult = geode::utils::numFromString<float>(yScaleStr);
+    auto rotationResult = geode::utils::numFromString<float>(rotationStr);
+    auto zOrderResult = geode::utils::numFromString<int>(zOrderStr);
+    auto layerResult = geode::utils::numFromString<short>(layerStr);
 
-        if(layer < 0 || layer > 32767) throw std::out_of_range("Layer is out of range.");
+    std::string errorString;
+    checkResult(xPosResult, errorString, "xPos");
+    checkResult(yPosResult, errorString, "yPos");
+    checkResult(xScaleResult, errorString, "xScale");
+    checkResult(yScaleResult, errorString, "yScale");
+    checkResult(rotationResult, errorString, "Rotation");
+    checkResult(zOrderResult, errorString, "zOrder");
+    checkResult(layerResult, errorString, "Layer");
 
-        if(std::isnan(xPos) || std::isinf(xPos) ||
-            std::isnan(yPos) || std::isinf(yPos) ||
-            std::isnan(xScale) || std::isinf(xScale) ||
-            std::isnan(yScale) || std::isinf(yScale) ||
-            std::isnan(rotation) || std::isinf(rotation) //i hate this
-        ) {
-            FLAlertLayer::create("Error", "Some of the values are either <cr>nan</c> or <cr>inf</c>", "Ok")->show();
-            return;
-        }
-
-        selectedObj->setPosition({xPos, yPos});
-        selectedObj->setRScaleX(xScale);
-        selectedObj->m_scaleX = xScale;
-        selectedObj->setRScaleY(yScale);
-        selectedObj->m_scaleY = yScale;
-        selectedObj->setRotation(rotation);
-        selectedObj->m_zOrder = zOrder;
-        selectedObj->m_editorLayer = layer;
-    } catch (const std::out_of_range& e) {
-        FLAlertLayer::create("Error", "Some of the values are either too large or too small.", "Ok")->show();
+    /*if(xPosResult.isErr() || yPosResult.isErr() || xScaleResult.isErr() || yScaleResult.isErr() || rotationResult.isErr() || zOrderResult.isErr() || layerResult.isErr()) {
+        FLAlertLayer::create("Error", "Error setting some of the fields", "Ok")->show();
         return;
-    } catch(const std::invalid_argument& e) {
-        FLAlertLayer::create("Error", "Some of the values is not a number.", "Ok")->show();
-        return;
-    } catch(...) {
-        FLAlertLayer::create("Error", "Unknown error.", "Ok")->show(); 
+    }*/
+
+    if(!errorString.empty()) {
+        FLAlertLayer::create("Error", errorString, "Ok")->show();
         return;
     }
+
+    selectedObj->m_positionX = xPosResult.unwrap();
+    selectedObj->m_positionY = yPosResult.unwrap();
+    selectedObj->m_scaleX = xScaleResult.unwrap();
+    selectedObj->m_scaleY = yScaleResult.unwrap(); //"real" hitbox??
+    selectedObj->setScaleX(xScaleResult.unwrap());
+    selectedObj->setScaleY(yScaleResult.unwrap()); //visual (?
+    selectedObj->setRotation(rotationResult.unwrap());
+    selectedObj->m_zOrder = zOrderResult.unwrap();
+    selectedObj->m_editorLayer = layerResult.unwrap();
+
 }
 
-bool ObjectEditorPopup::setup() {
+bool ObjectEditorPopup::init() {
+
+    if(!geode::Popup::init(485.f, 230.f)) {
+        return false;
+    }
 
     //if you are seeing this and have any suggestions, please let me know. i hate cocos ui
     m_noElasticity = true;
@@ -141,7 +153,7 @@ bool ObjectEditorPopup::setup() {
 
 ObjectEditorPopup* ObjectEditorPopup::create() {
     auto ret = new ObjectEditorPopup();
-    if (ret->initAnchored(485.f, 230.f)) {
+    if (ret->init()) {
         ret->autorelease();
         return ret;
     }