+++ changelog.md
@@ -1,2 +1,8 @@
+# 1.0.1
+- Ported to Geometry Dash 2.2081.
+- Fixed error with scale.
+- Fixed wrong type in xpos and ypos field.
+- Better error messages.
+- Internal changes.
 # 1.0.0
 - Release